Goddamn it state machine screwed me over
The example code never showed the part where the client state was set to
hello, yet the state machine was continually checking for it.
The state machien now upgrades connected states to hello, so it can
continue. Hello and Connected seem redundant, maybe i just get rid of
the hello.. gotta check other areas of the code though.